added my 4 modules from the demo
